wip: upgrading sveltekit for adapter static
