+++ src/app.html
@@ -40,9 +40,10 @@
 
     <!-- This contains the contents of the <svelte:head> component, if
 	     the current page has one -->
-    %svelte.head%
+    %sveltekit.head%
   </head>
+
   <body>
-    %svelte.body%
+    %sveltekit.body%
   </body>
 </html>
